streamlit_app/requirements.txt
streamlit==1.29.0
chromadb==0.4.18
ollama==0.1.7
langchain==0.0.350
langchain-community
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.4
tiktoken==0.5.2
